+++ packages/components/src/local/coa-status-board/helpers/gen-data.tsx
